backend/agent/requirements.txt
langgraph>=0.0.60
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
openai>=1.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
